Cool findings:
We probably want to have an io adapter similar to gzip_stdio_adapter that
handles cache files.

We need to make reset (or a prepare to begin reading) a failable operation in order to check whether cache_numbits follows
the right size rule.

Either that or unify how we handle bad file headers between gzip and cache files.